refactor: foreign chain impl block and test utils into separate package
