+++ crates/contract/src/api/test_utils.rs
@@ -11,7 +11,6 @@ use crate::primitives::thresholds::{GovernanceThreshold, GovernanceThresholdPara
 use crate::state::ProtocolContractState;
 use crate::state::test_utils::gen_running_state;
 use crate::storage_keys::StorageKey;
-use crate::tee::test_utils::Environment;
 use dtos::{
     Attestation, Curve, DomainConfig, DomainId, DomainPurpose, MockAttestation, Protocol,
     ReconstructionThreshold,
@@ -19,13 +18,11 @@ use dtos::{
 use elliptic_curve::{Field as _, Group};
 use k256::elliptic_curve;
 use near_account_id::AccountId;
-use near_mpc_bounded_collections::NonEmptyBTreeMap;
 use near_mpc_contract_interface::types as dtos;
 use near_sdk::store::{IterableMap, Lazy, LookupMap};
 use near_sdk::test_utils::VMContextBuilder;
 use near_sdk::{NearToken, VMContext, testing_env};
 use rand_core::CryptoRngCore;
-use std::collections::BTreeSet;
 use std::str::FromStr;
 use threshold_signatures::confidential_key_derivation as ckd;
 use threshold_signatures::frost_core::Group as _;
@@ -262,53 +259,6 @@ pub(crate) fn forwarded_participant_call_contract() -> MpcContract {
     testing_env!(ctx);
 
     contract
-}
-
-/// Votes `chain` into the on-chain RPC whitelist with the signing threshold of participants.
-pub(crate) fn whitelist_chain(contract: &mut MpcContract, chain: dtos::ForeignChain) {
-    let batch = NonEmptyBTreeMap::new(chain, ::test_utils::contract_types::dummy_chain_entry());
-    let threshold = contract.threshold().unwrap().value() as usize;
-    let participants = participant_account_ids(contract);
-    assert!(
-        participants.len() >= threshold,
-        "need at least {threshold} participants to whitelist a chain, got {}",
-        participants.len()
-    );
-    for account_id in participants.iter().take(threshold) {
-        testing_env!(
-            VMContextBuilder::new()
-                .signer_account_id(account_id.clone())
-                .predecessor_account_id(account_id.clone())
-                .build()
-        );
-        contract
-            .vote_update_foreign_chain_providers(batch.clone())
-            .expect("vote should succeed");
-    }
-}
-
-pub(crate) fn register_foreign_chains_config_for(
-    contract: &mut MpcContract,
-    account_id: &AccountId,
-    chains: impl IntoIterator<Item = dtos::ForeignChain>,
-) {
-    let foreign_chains_config: dtos::ForeignChainsConfig =
-        chains.into_iter().collect::<BTreeSet<_>>().into();
-    // In mock setup, account_public_key == tls_public_key.
-    let tls_key = contract
-        .protocol_state
-        .threshold_parameters()
-        .unwrap()
-        .participants()
-        .info(account_id)
-        .expect("account must be a participant")
-        .tls_public_key
-        .clone();
-    let mut env = Environment::new(None, Some(account_id.clone()), None);
-    env.set_pk(near_sdk::PublicKey::from(tls_key));
-    contract
-        .register_foreign_chains_config(foreign_chains_config)
-        .expect("register should succeed");
 }
 
 impl MpcContract {
